refactor(engines/pi): activation needs no lock, so it needs no second copy
Activating a deferred tool is a read-modify-write against pi's active set, and
every step of it is synchronous. An async `activate` could not say that, so it
carried a promise chain to serialize what cannot interleave — and the chain
pushed an await boundary out through `defineTool`, leaving the invariant
observable only by comparing function identities.

`ToolActivation.activate` answers `string[]`. The bridge is one function in
tool-context.ts, beside the contract it implements: serving passes the
`onActivated` callback that records the delta, chat has nowhere to put one, and
that parameter is the whole difference the two copies used to encode.

The turn context is built once per session rather than per tool call, because
it describes the session — a tool call has to see what the previous one
activated.
